A partitioning loader that tracks the origin of a configuration key.
Key changes:

 -  Refactor loaders (and their tests) into packages for clarity
 -  Introduce `PartitioningLoader` which:
     -  Loads from several other loaders (think: `load_each`)
     -  Keeps a record of which values come from which loader
 -  Introduce `load_partitioned` and `load_config_and_secrets` syntactic sugar
